Lots of bug fixes.
Especially with constructing the train track maps. Now in the case of
the genus 4 non-orientable Arnoux-Yoccoz, the three Pseudo-Anosovs and
depth one are found, but they are not recognized as PA's yet, because
the matrix is not primitive. Interestingly the largest eigenvalue is
already what it should be, but there is probably still a bug, because
the transition matrix of a train track of a PA should be primitive.