Added support for more general eigendata calculation.
So far only eigendata of PF matrices was supported, now it is determined
for any matrix if is has a chance of being a factor of a PF map, by
checking the dominant eigenvalue and eigenvectors.

This is useful, because the matrices the come up naturally are not
PF. Even the transition map of train tracks is not in general, because
of the 2-pronged vertices.