Implemented the search in a train track complex.
Overall, the algorithm seems to be working. Building the complex is
currently very slow, mainly because of the inefficient train track
transformations (implemented by transformations of foliations which use
the whole transition_map.py stuff which is not necessary).

Also, the Teichmuller polynomial calculation seems to be buggy in
certain cases. The invariant cohomology calculation is likely to be one reason.